REFACTOR: reorganize tool display components
