+++ src/components/tool_displays/GripperDisplay.vue
