+++ src/components/tool_displays/RobotHandDisplay.vue
